:hammer: refactor: parseConfig -> config로 폴더 이름 변경
